rename to lib + add defs
