+++ workflows/lib/defs.ts
@@ -1,0 +1,14 @@
+/**
+ * Definitions used for the other workflow modules.
+ * @module
+ */
+
+/**
+ * List of packages in the monorepo.
+ */
+export const PACKAGES = [
+    "lenskit",
+    "lenskit-funksvd",
+    "lenskit-implicit",
+    "lenskit-hpf",
+];
